[2/2] Rename ElasticsearchService to QueryService
Rename module
