+++ src/azul/plugins/metadata/anvil/service/aggregation.py
@@ -8,7 +8,7 @@ from opensearchpy.helpers.aggs import (
 from azul.plugins import (
     FieldPath,
 )
-from azul.service.elasticsearch_service import (
+from azul.service.query_service import (
     AggregationStage,
 )
 from azul.types import (
