+++ src/azul/plugins/metadata/anvil/service/response.py
@@ -25,7 +25,7 @@ from azul.json import (
 from azul.plugins import (
     SpecialFields,
 )
-from azul.service.elasticsearch_service import (
+from azul.service.query_service import (
     ResponseTriple,
 )
 from azul.service.repository_service import (
