+++ src/azul/service/index_controller.py
@@ -39,13 +39,13 @@ from azul.service.controller import (
     validate_catalog,
     validate_params,
 )
-from azul.service.elasticsearch_service import (
+from azul.service.query_controller import (
+    QueryController,
+)
+from azul.service.query_service import (
     IndexNotFoundError,
     Pagination,
     SortKey,
-)
-from azul.service.query_controller import (
-    QueryController,
 )
 from azul.service.repository_service import (
     EntityNotFoundError,
